gemini-debate/v3/requirements.txt
google-generativeai
pyyaml
